[FIX] hr_timesheet: add percentage formatter for project_task_progressbar
Before this commit, the total for each group or the grand total
displayed in the list view of task for the progress field is formatted
as a float instead of a percentage. The reason is because the framework
does not find the formatter associated to that widget and so we will
take the default formatter for float field.

This commit adds a formatter for the widget to be sure the framework
formats the total displayed in the list view in percentage.

closes odoo/odoo#201157

X-original-commit: d7551871917e44a709324c21103e2fd18ba81398
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>